SNBT Improved Array Parsing Errors
This actually worked out really well! Maybe I don't need to take the functional approach to organize things, the parameters still work great here too. I'm still up to trying to figure out that other approach though.

To make the error show up in the test runner, you can make `listItemAssertion ` set to `false`, and it will show the error for trying to parse the broken file associated with that check.
I'm still not totally sold on what to call the key for the RootTag, in the event that the root tag is indeed a ListTag.

I still need to add this check, as well as these improved errors, to `#readArray()`, where each list item type can be validated as the correct type or not.

#39

My eyes are getting so heavy/sleepy now, nearly done with these for the night.