TS 5.3!
Yay!
Unfortunately, it didn't fix my issue with needing to cast tag values with 'as' when referencing them in between their relation to their tag types. I was hoping that TS's stronger syntax with the 'switch' handling in this update could help with that (after expanding the return type associations for 'getTagType()', but it still didn't work)