`TAG` Enum
Alright, I think I'm up for trying it out now! Enums seem pretty useful, and I think this may be more ergonomic for the users of NBTify, in the event that they need to use part of the format's implementation spec.

Trying to keep up with coding, it's been hard now that I've been back to work again. It's starting to get busy for the holidays!!! I haven't been able to run this week yet either, which I want to try my best to fix maybe. With my work schedule and the short daylight, it's not feasible to get a run in during the daylight hours. That's been the biggest hangup. I want to run, the time is the issue.