Node.js Deflate Raw
Node now supports `deflate-raw` for the Compression Stream API, as of version `21.2.0`.

Because of this, I'm going to remove my polyfill in NBTify to handle using the `node:zlib` API, and instead fully use the Compression Stream API. I feel like the better way is to just leave the polyfill in since this was only added about 10 days ago into Node itself, but I'm feeling good so I'm going to move things forward now. If worst comes to worst, I can remerge the polyfill back in. Barely any NBT implementations use `deflate-raw` as it is (only LCE files to my knowledge, let alone for plain NBT that NBTify handles), so this isn't something that will come up very often, as it has to be a rare `deflate-raw` NBT file, and someone would be using NBTify in Node.js in not the latest version. Most people will be using NBTify by way of Dovetail, which is in the browser (essentially fully supported there anyways), and I also use a polyfill there just in case. I think NBTify should be lean enough to not use many polyfills or fallbacks for it's own implementation. If anything, the user should handle that sort of thing, like I am doing with Dovetail :D

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Object/defineProperty
https://developer.mozilla.org/en-US/docs/Web/API/Compression_Streams_API#browser_compatibility
https://github.com/mdn/browser-compat-data/issues/21343
https://github.com/nodejs/node/blob/main/doc/changelogs/CHANGELOG_V21.md#2023-11-14-version-2120-current-targos
https://github.com/Offroaders123/NBTify/issues/9
https://github.com/Offroaders123/NBTify/issues/11