Stricter SNBT + TS 5.3 + Node Deflate Raw
And I also started working on an initial experimental implementation for VarInt little endian encoding, but it wasn't working out correctly, so I'm holding that back to a later release, and I will try again at that point to reimplement it again. I think the challenge is understanding how the implementation works in regards to how it lines up with the NBT spec, and where VarInt is specifically used in the NBT implementation.

Oh, and I forgot to mention in the last commit that I made the change to the `Object.defineProperty()` calls for the `NBTData` fields because I wanted to more explicitly show how I want them to be non-writable, and I was just relying on the implicit behavior of having them be non-writable, which is what happens when you don't pass in the `writable` flag to `Object.defineProperty()`. I didn't add it originally because that was the default, so it made sense not to need to pass it. But after thinking about it, showing that it's moreso intended behavior by clarifying it in the config is more simple to see, rather than needing to read the docs like I did, to know that it does that implicitly.

In other words, with that and my TS variable and function return types, I want to start being more explicit once I know how I want things to work a certain way. It helps make things easier to catch unexpected changes down the road, because my pre-defined breakpoints show me where my intended behavior meets my unexpected changes, unlike having it bubble up to the nearest type check. That example is more in regards to return types, but yeah!

I want to try using more spaces in my code formatting too, for the longest time I have been used to doing no extra whitespace, not completely sure why. I think I was just used to that from the very early days, and I just kept that methodology around for a long time hehe. My opinionated formatting is what initially turned me off about Rust, since they have a very strict set of formatting rules, and I want to use Rust, so I want to get myself out of the habit of needing to use the syntax formatting that I personally like, and learn to like the other ways also, since I don't want that to hold back my learning and experience.