Schematic, Litematica, Structure Block, Infinity
Right now the tests aren't passing, and that's because of my current handling for ListTag item types, for empty lists. If a ListTag is empty, right now NBTify will just type the list with an EndTag as the item type. So to make things compile symmetrically, I need to find a way to persist the intended item type for empty ListTags.

https://www.reddit.com/r/technicalminecraft/comments/gf62x4/how_to_convert_litematic_file_to_schematic_file/
https://github.com/GoldenDelicios/Lite2Edit
https://minecraft.wiki/w/Schematic_file_format
https://github.com/SmylerMC/litemapy
https://github.com/maruohon/litematica
https://github.com/Lunatrius/Schematica
https://minecraft.wiki/w/Structure_file
https://learn.microsoft.com/en-us/minecraft/creator/documents/introductiontostructureblocks?view=minecraft-bedrock-stable
https://github.com/microsoft/minecraft-samples/blob/main/structure_blocks_sample_behavior_pack/README.md
https://www.youtube.com/watch?v=fQZOPAoxw9Q
https://www.planetminecraft.com/project/small-medieval-house-lobby/

Devin Townsend Infinity, I'm loving the sound of this one. The story behind it is very driving too.

#40
#41

I really need to work on an improved organizational setup for file type subtests, right now they are all grouped together into a single one, and it's hard to determine what is doing what now. Originally it wasn't so bad because I was only testing against binary NBT files, but now I have things like these world save types, SNBT, and things like that. Each one needs a little bit different of handling for how it should be tested.

Another important thing I noticed is that I tried to open a fairly large `.schem` file, and it crashed Dovetail, either because of the UI, or because JS can't handle it's decompressed size, or it's parsed NBT object size, something along those lines. So it afterall may be too hard on the JS implementation of mine to try to do heavy lifting work on world data if it's having trouble opening a Schematic file. I'm hoping I can manage it effectively, but I'm not totally sure. For this big project to work, I may *need* to move to another language for it to work. Up until now, I haven't had any issues with NBTify's power to parse and save things, but it is concerning to have run into them now. Part of the song I'm listening to sounds like it's saying 'don't give up' though, right as I'm writing this, so maybe I should keep trying anyways though :P