SNBT Empty ListTag Types
Check out the message about my findings for this commit, in #40 down below, above when this was referenced in that feed when you click on it.

Note how the tests are still not passing, it's because SNBT has no way of tracking the types for empty ListTag entries.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Symbol
https://stackoverflow.com/questions/31897015/what-is-global-symbol-registry
https://stackoverflow.com/questions/49619054/whats-the-purpose-of-registry-symbols-symbol-for-in-javascript-es6

#40
#41

And woah! Who knew the fonts for road signs are standardized! The main one is called Highway Gothic, and it's been around since the 50s. The new one was going to be Clearview, but looks like they reversed making that one the new standard. It's present across a lot of the southwest, I see it a lot when crossing the river into Arizona. I really like it! I hope it stays around.