Recompile Check
This is a recompiled version of the Schematic file, I'm going to use this for the proper diffing of the file for the test. The only changes to the actual output having used it in NBTify is the gzip compression format layout, and the empty ListTag type instance. I'm going to manage that on a check on it's own, outside the scope of this file specifically.

Ants is a good song I think, it's my kind of jam for sure. I really like how many places this album goes. I'll have to check out Ocean Machine at some point too, he mentions that one a good deal, it seems to be one of his more widely known early projects.

#40
#41