ReadOptions Spread
This is more maintainable, in that now all of the `read()` options are now passed down to their child calls automatically. This wasn't type-checked with my current setup yet, so a few times I had missed passing down newly added options to sub-recursive checking calls. I am also going to look into adding `satisfies` statements with these options objects, to ensure that all of the appropriate values are being passed in at various stages of the function. That may be a little much though, I think this should do.