RootName + Bedrock Level Primitive
Changing the naming of `NBTData.name` to `rootName`, as well as it's type to `RootName` from `Name`. This is to further distinguish what this feature is for.

Changing the value type used for `NBTData.bedrockLevel` to a plain `number` from `Int32`. I think I originally used `Int32` because I wanted to ensure that the user passes in a number valid to the `uint32` range, but using `Int32` for this is a little counterintuitive because of the signed implementation difference. Maybe I can look into a `Uint32` wrapper class implementation, but I don't much see a use for that as of yet, since that's the only public place this is used in NBTify. When reading and writing values for the Bedrock Level from the file buffer, it already ensures that it's within the correct range, by way of wrapping. So I think the decision here is as whether to throw an error for this range detection, or if it should wrap it to fit in the given space. Say if it's bigger than the largest value for `uint32`, it will wrap back around to `0`. This number isn't supposed to be big at all though, since it's just for the version of the file. So this isn't an issue that should likely come up until there are a LOT of Minecraft versions. That will be an issue for Mojang itself as well, since they are relying on the same format I'm implementing. (I'm implementing their format)