MUTF-8 Support
This is just a temporary thing to test things out. I need to add another test which includes MUTF-8-specific strings into an NBT file, because right now all of my tests are passing either way, even though NBTify wasn't MUTF-8 compliant, meaning that it's not 1:1 in serialization with what's expected in the NBT spec.

And this part is temporary because I'm debating whether I should bundle `mutf-8` into and start building NBTify for the browser, rather than going the no-dependency route where I only transpile with TS. If I bundle for the browser (ESM), then I can use regular dependencies in Node, and put everything together for use elsewhere. If I do that, maybe I will rethink CJS support as well, since it's not that hard to simply enable nowadays. I've wanted to try building/transpiling with plain ESBuild for some time now as well, since it will allow for a minified output for the web, meaning you can still get tree-shaking when installing locally and using with bundlers, but you can use it feature-complete in the browser as well.

https://wiki.vg/NBT#Specification

#42

https://github.com/sciencesakura/mutf-8/pull/17