Un-Demo MUTF-8
The previous commit was just for demo's sake, to try things out. Going to come up with a proper solution for that down the road.

Starting to check out and listen to Boil That Dust Speck. The darker heaviness with the lower tunings caugnt me off-guard for Mike's usual aesthetic, I already really like the sound of it! Starting with a little set of songs at a time.