Empty List Type Persistence!
Yay, it's working!

Was going to use `Symbol.for()` to implement this, but I think I'm going to use NBTify's exports for it instead. Since you are likely going to import features from NBTify to read and write NBT files anyways, I think it makes more sense to import the Symbol for List tags from there if you needed to use it for something.

https://stackoverflow.com/questions/34587901/difference-between-symbol-and-symbol-for
https://nodejs.org/api/util.html#utilinspectcustom
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Symbol

Notes on how this works for the library itself:

Going to use an optional Symbol() property to keep track of the type on the array

You don't have to define it manually yourself when writing NBT files from scratch, this is mostly meant just to ensure existing files like these can retain their existing type information
When reading files from NBT files, it will always attach this property to your `ListTag` arrays, and it will use them in the writing stage if they are present
I want them to be optional because I still want you to build your own NBT shapes using plain arrays, and not need to assign that property yourself every time

(This is from Discord DMs, so that's why this is a little weird of context)
Ooooh
I bet I can define that as a hidden property too, so it won't clog up things when logging it to the console
Of course when you manually add it yourself like that it will show up, but from the reader it could do that for you

(Back to normal commit discussion)

As mentioned in the commit "SNBT Empty ListTag Types", SNBT doesn't have a viable way to persist this information, so this is an unfortunate case where going back and forth from SNBT with empty list tags will lose the original item type. If you have any ideas on how this could be handled, that would be great! I think Bedrock's lack of SNBT support is more of an example/correlation as to why their parsing code requires specific List item types, because if they supported SNBT in commands and such, they would need to allow un-typed (End tag) empty List tags as well, since they would have the same problem there with SNBT, since the format itself just doesn't have a way to represent the item type for empty lists. I guess the difference between my use and the game's implementation though is that theirs is likely driven by runtime types, so it gets validated or checked there instead. I'm not sure. Still a bit funky that it may not support un-typed empty Lists, a bit peculiar indeed.

#40

rBnaodn "so, many, mistakes" (Me!!! Started working on releasing it just last week or so)
Marco Minnemann "My Sister" (Remembering how much I love this album, and just how detailed it is!)