Expanded Input Types (Concept)
While working with NBTify in updates to Dovetail, I've been debating whether it may be useful to allow more than Uint8Array and ArrayBufferLike objects into the NBT reader functions. I'm still not sure it's much more helpful, but it's still a concept. I decided to try that out here.

I also tried a concept of making an "NBTFile" builder class, with methods of things like `.blob()`, `.arrayBuffer()`, `.littleEndian()`, and other things like that. I'm not sure how I feel about that though. It seems to high-level for something that's already kind of trivial with NBTify's current setup. It looks nice, and it seems really fancy, but I don't think it's actually worth it, because it's not doing anything crazy at the runtime level, only at the reading and writing stages. I may still delve into it as a concept though, to see what comes of it. Essentially I wanted to see if I could morph/make a companion for `NBTData` so it allows more features to that like `Blob` or `File`.

I'm also thinking of making the properties on `NBTData` mutable, since they only weren't previously because of assignment validation concerns. I think that's not too much of a problem though, so maybe I just will add that. It eliminates a lot of the hairyness around type checking, since it will all be scoped to that single class, rather than needing to be put all over the entry points of the NBTify codebase. I still have to do this for other parameters and things though, say like for `Uint8Array`s and things like that. I think I could more generalize where errors are handled though, so I don't have to validate all over the place.

Not sure how I'm going to go about it yet, but with the things I've learned lately with scoping, over in the new Dovetail codebase, I think I also want to move NBTify's reader/writer implementations away from classes, and into standalone generalized functions.

https://developer.mozilla.org/en-US/docs/Web/API/Response/Response
https://wiki.bedrock.dev/nbt/step-by-step-example.html