Proxy Testing
So close! This works really great, but logging the object to the console shows the Proxy and the target object itself, so it's not completely hidden away. I guess I will stick with my constructor-based validation. I may still keep the separate functions I made here though, those can better be used in other places throughout the code as well.

I see how this is less of a problem for frameworks and things like that though, where you don't usually log plain values to the console. Maybe that could be a simply way to achieve nice validation for things though, I will absolutely keep this in the back of my mind for things, down the line. I'm curious if this is kind of how Zod works somewhat too? I know the API for that is structured a bit differently, but yeah.

https://stackoverflow.com/questions/37714787/can-i-extend-proxy-with-an-es2015-class (back to this one, same from the last commit)
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Proxy/Proxy/set
https://stackoverflow.com/questions/61765782/javascript-proxy-how-to-hide-target-and-handler-in-console-and-debug