Hybrid Constructor Validation
I have been wanting to have something like this setup for some time now, I think I will try this again a little later, moving this style to other parts of the codebase as well.

Having a loop for the property definitions is so much more simple now as well! I wanted to do that a while back, but I don't think I fully understood a way to do that just yet.

This brings the library's behavior back to before the demo validation things. The wider-supported IO parameters is still in the codebase as well, going to see whether that should still stick around as well.