NBT Error Parse Info
Working on figuring out the concepts behind making your own error handlers, in that I can provide the missing metadata I need to be able to get without needing to parse the error message itself.

This is a big of a weird setup at the moment, since I don't quite know the correct way to structure these kinds of extensions.

https://developer.mozilla.org/en-US/docs/Web/API/DOMException
https://stackoverflow.com/questions/5136727/manually-artificially-throwing-a-domexception-with-javascript
https://javascript.info/custom-errors
https://stackoverflow.com/questions/67558074/how-to-throw-custom-errors
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Errors
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Error/cause

#31