BedrockLevel Boolean - Derived Version
Adds full support for deriving the BedrockLevel header version value from the `StorageVersion` key. Less data-duplication is the way to go!

https://github.com/Offroaders123/PiPe-Levels/commit/abd543c6b85d9f597d19993bfd717ad5c4ebaca3

I am planning on refactoring NBTify pretty much as a whole, to hopefully/ideally being defined as standalone functional 'components' of some sort, rather than using builder-class kind of things. At least for the SNBT module, I want to have that. NBT though, it may make sense to still have a builder class. I think the `read()` function should be much less verbose though, I want to move a lot of the logic out of there, and break things up a bit. Most of the library could use that in a ton of different spots. I'm still not completely conceptualized on how this will look yet, but I know the current one isn't up to par for me anymore.