BedrockLevel CLI Assignment Validation
Now the CLI won't allow you to add the `--bedrock-level` flag if the input file doesn't have a `StorageVersion` key to use it's value from. It already checks this when writing to an NBT file, but I want it to let the user know sooner in the process so they don't have to write the file in order to find out that the flag is invalid for the file they are formatting.