Dual ESM/CJS Demo (Concept)
Looking into whether this is something that NBTify would actually be neat in having. I'm not completely sure it's necessary yet. (In terms of what I think this package is meant to be used for. No doubt do I think dual-build packages are amazing in their own right!)

I think I'm trying to decide whether the complexity of supporting both is worth the hassle in terms of it splintering into multiple ways of potentially using the package. I don't know if I fully like that concept. It reminds me of the divide between Java and Bedrock as well.

I'm not sure about the bundle-related things yet either, from a package perspective. Should I bundle dependencies for use in Node? Should it be minified? If things are bundled for the browser anyways, people have mentioned that you should let the users of your package decide how they want to handle the build of your package, not to have you do it for them. I can understand that as well.

https://antfu.me/posts/publish-esm-and-cjs
https://medium.com/swlh/npm-new-package-json-exports-field-1a7d1f489ccf
https://nodejs.org/api/packages.html#package-entry-points
https://nodejs.org/api/packages.html#dual-commonjses-module-packages