SNBT Exponential Doubles
Realized this was an issue, when converting back and forth from SNBT, to JS, as well as writing back to SNBT. Now the writer properly handles these exponents/decimal point handling as well, for these additional circumstances.

The issue with the writer is that it was adding `.0` to the end of a Double tag, when it already had a `.` at the start of the scientific notation for the number value, meaning it was writing it like `5.252e+5.0d`, which doesn't make any sense. This would then be parsed as a string, leading it to be invalidly parsed when part of a List tag, where only tag types of Double would be allowed.

This helps me also see how it could be helpful to use this behavior to further help assert that all variants of a given tag are parsed and written as the same type, where you could write each tag's test case as part of that List tag, and it would be an assertion that each one is read/written as the same types that surround it. Could be great! I guess you could also just write a JS test to assert that, which is also likely the more predicable way to explain that functionality.

https://stackoverflow.com/questions/55724674/how-to-check-for-exponential-values-in-javascript
https://stackoverflow.com/questions/70773991/how-can-i-check-if-a-number-is-in-scientific-notation-with-exponential-part
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number/toExponential

#45