JSON File Support! (CLI)
Now you can simply convert between NBT, SNBT, and JSON, just with NBTify in the CLI!

This is inherently not meant to be a lossless conversion step, but rather one that is just a simple helper for maybe easier data visualizations, as well as just wanting to view a given format as another kind.

#25

Wow, cannot believe I've been working on the CLI for a year now! Or at least exploring it. That means I've about had my Mac for a year now coming up soon as well, that's crazy. This last year has been my most productive one yet. I only want to keep trying to see what I can further accomplish and improve on from here. It's a combination of both getting more focused on accomplishing things, as well as letting go when it's not the right time to work on a given thing, learning to give myself a break when I need one.