Further Class Rehash Experiments
This was from yesterday, I kind of left off to work on other projects. I think I may stick with the class implementation for the underlying structure of the state handling, it's a bit easier to use that, than to use strictly only functional APIs to build and manage the state of the reader. I think the scope of the classes are what need to be shrunk, maybe not to remove them completely.

I am thinking about removing/abstracting out the NBT implementation's specific use of buffer management, into their own classes, this is similar to what Evergreen and Jerrin do, maybe Prismarine-NBT as well. It's fine how it works now, in the technical sense, but I'm realizing this buffer read/write handling could be helpful to make use of in other situations as well, one example being LCE chunk data, which isn't NBT. So I need some sort of tape feeder to read that as well, but I don't want to write a whole new buffer implementation to do that. I think NBTify could break apart what is buffer-related, versus what is NBT related. Come to think of it, that may actually remove the issue of needing an NBT-specific class as well, since the reading state logic can be used for that buffer class instead, and the NBT part doesn't need to care about the internal state for that. Going to look into that! :D