Hello World Parsing!
Got `hello_world.nbt` parsing all working with the tape reader, as well as the standalone NBT functions! This is great. So the buffer-centric code is now separate from the NBT reader implementations. I think this could make things more doable as it moves along. This will make handling for file-specific headers much simpler to implement as well, which is great.

MacAlpine - Equilibrium