Most Tag Writing
Mostly working, still a few inconsistencies to iron out.

It was actually "Genetic Drift" that sounded like Terminal Velocity to me, like that album in general. Really love TM and JP! I want to see an album with TM and Mike Keneally too.

Yeah ran Diffchecker on a stringified version of the bytes from the buffers, and it isn't identical yet. Going to take a break to eat, then come back and debug what's going on here. So happy how nicely these are working out!