Write Tag Writing
Now it fully parses, and builds `bigtest.nbt` symmetrically! I found out what the issue was with the codebase from earlier. It was a small mistype for writing `Float64` values (was accidentally set to `Int16`, I think from when I was copy-pasting the values), the length written to strings didn't match the buffered version of the string (the length of a string's characters doesn't always line up with the amount of bytes when written to a buffer, think of like how special UTF-8 characters are encoded with more than one specifier), and I'm not sure if it was causing part of the issue, but there was an inconsistency for the `allocate()` function validation check, which was using the wrong property to check. Oh, and the extra bytes not being trimmed was because I didn't copy that feature over to the new reader yet, I thought `allocate()` did that for some reason. I need to figure out what `allocate(0)` does, I don't think it does anything? There was also an extra unexpected use of a `littleEndian` parameter where it wasn't intended to be passed in.