BedrockLevel Support
Realized that this was part of what I wanted to do for a while now, adding the BedrockLevel handling into the main part of the parser, rather than having it be part of the test function itself. I want the standard set of NBT features to be available from the regular base APIs, not only the detection-related ones.

I think I realized with this one where I need to distinguish between the write/read pointer, versus the completion byte offset counter. For example, I tried to move the 'pointer' for the writer back to the start of the file, then write the BedrockLevel header to the start of the file. This made it so `trimmedEnd()` would remove the entire rest of the file though, because now the pointer only shows up to the BedrockLevel header. I was having this issue with the LCE parsing as well, where I want to move the pointer to a specific place, but don't want to reset the write completion offset. I'm going to make note of this, and it will be much easier to get LCE's parsing set up, since the byte-related things are no longer implemented to be NBT-specific. That's one of the main goals with this rework, I want to have more universal tape reader buffer implementations, which can be used for general reading and writing, regardless of whether it's to work with NBT. Now I see how/why Prismarine-NBT seems to do this, I think this is most exactly what their ProtoBuf package does.

Oh yeah, and this is tested on the `level.dat` file now.