ReadOptions Parameter
Consolidated the reading options into an object, to match that of the NBTData-like kind of object.

I want to further look into what might be a better way to provide a writing API that's a bit more straightforward, because the use of both an NBTData object in conjunction with a FormatOptions object is a bit confusing.