NBT Usage Generics + TagType Function
I also removed the return types from the writer functions, I'm not sure I want to expose being able to get their raw buffers or not, but we'll see. That was kind of the purpose of doing things with this functional setup. Maybe the `readTag()`/`writeTag()` functions can instead be what returns the buffer data, instead of the individual implementation functions. I think I may use that system, since it further less de-couples complex parameter usage and API complexities.

Essentially, `readTag()` and `writeTag()` could be to work with any arbitrary root tag types, while the plain (already existing) `read()` and `write()` functions are for regular use NBT types, where it's either a root list or compound expected. Essentially, I want the custom root usage to be opt-in at the API level, not something that just happens by default. That's also part of what makes the auto-detection functional and sound, because it only allows a set range of factors to be possibly parsed correctly.

I'm also considering about making `NBTReader` and `NBTWriter` classes again too, but that may be going backwards for all of the work I figured out for this so far.