Test Setup
Modifying the current build so it can more-properly run most of the tests, they are passing partially! So the next thing to look into is format-auto-detection, and usage of NBTData format parameters.

Realized I forgot to link the SNBT modules up to the new class/functions references, so things like `instanceof` weren't working correctly, because they were referencing the wrong declarations.