API Smoothing
Trying to implement a few other things. I want to make NBTData non-readonly anymore, and I want to make the validation be consistently applied only at the read/write stages, and at the type-level. NBTData should be a mutable shape, with editable keys/fields. While it does allow for odd things when using that shape in the rest of the API, this can happen for other things across NBTify anyways, like if you try to pass in a goofy object shape into `new NBTData()`. I don't think the validation should happen at the class, and it should only be when it's used, which is how the rest of other NBTify concepts work as well. Another example of that is the usage of `ListTag` for example.