Tag Aliases Support
Inspired in a concept from nbt-ts, I think it makes sense for NBTify to be able to handle other JS focused objects/classes, like the various options for the TypedArray classes. Rather than writing a `Uint8Array` object to NBT as a CompoundTag, it can instead just coalesce it into being an `Int8Array` of sorts. It makes more sense to do that storage-wise, and it's a bit more straightforwards to account for as well. This helps add better support for other subclasses as well, like `Buffer` in Node.js. Come to think of it, should I also account for the `ArrayBufferLike` classes, and wrap them with an `Int8Array`/`Uint8Array` as well? I can look into that later. I think part of also wanting to do this was my experimentation with writing an entire Bedrock world's LevelDB contents to a single NBT file. However, the use of `Buffer` objects for my current representation of not-yet-implemented database keys meant that they were inefficiently stored as CompoundTag keys, which adds much more data to write to the file, because each key would get it's own type, unlike how the TypedArray-based ones only have a single distinguished type to account for. Another thing is that the data is still saved safely, in that it's the same data being written, it's just how it's viewed after being re-read again that has to be explicitly handled if you expect the unsigned variants, which don't have explicit NBT tag types. So the buffer data itself is still lossless, it's just how you decide to read it, which is why I think it's nice for it to just work, in the writing sense of things.

I'm not sure I've outright mentioned all of these library inspirations for NBTify, but these are the main ones I have looked towards over time to get different concepts about implementing things in a certain way.

[NBT.js](https://github.com/sjmulder/nbt-js)
[Prismarine NBT](https://github.com/PrismarineJS/prismarine-nbt)
[nbt-ts](https://github.com/janispritzkau/nbt-ts)
[Anvil Parser](https://github.com/Geoige/anvil-parser/blob/master/nbtParser.js)
[MCNBT](https://github.com/BoogeeDoo/mcnbt/blob/master/nbt.js)
[Deepslate](https://github.com/misode/deepslate/tree/main/src/nbt)
[l2nbt.js](https://github.com/lgou2w/l2nbt.js/blob/develop/src/nbt.ts)
[deno-minecraft](https://github.com/janispritzkau/deno-minecraft/tree/main/nbt)
[objecttonbt](https://github.com/Jakobg1215/objecttonbt)
[l2mc.js](https://github.com/lgou2w/l2mc.js/tree/develop/packages/mcnbt/src)

https://www.youtube.com/watch?v=ea2VM1zuFdE&list=PLpA5ThqzxTYX3hbSyAVsxu8c9tb68-krF&index=15