Primitive Tag Generics + Demo Tidying
Kind of following the easy use of tag types in the previous commit, I'm also looking into the concept of making plain tag types generic as well, to ensure the definitions for NBT typings are consistently implemented, so they can be more statically analyzable. So rather than a typing for a specific tag being just `ThingResource`, instead it will be written as `StringTag<ThingResource>`. Since NBTify just uses JS primitives where possible, this doesn't represent anything different, than just a string enum or union type. So it's not functionally any different, but I think it can make it possibly easier to visually discern (or even derive other language types from) when you see what tag type is wrapping it. So rather than having to look at `ThingResource` to figure out what type that is, you already know ahead of time that it has to validate against `StringTag`, which means it has to be a string.

Now the 'rebuild' demo function can run with any of the other files, like how it was from before, with `process.argv`.