Flattened Reader-Writer Methods
I thought I wanted to go back to this setup, but maybe not, afterall. Going to still commit this for tracking it as a concept. I think I might instead want to use an enum to associate the byte lengths of each number type, to that number type. I think that can cleanly remove some of the ambiguity overhead with implementing each of these like this.

```
I think the last part he mentions, is the concept I'm trying to go with for NBTify, where it's API is meant to provide you the building blocks to implement it's feature set into your own codebase's needs, rather than your own codebase to be built around the API itself
https://www.youtube.com/watch?v=CXyzmS2eNFo
```

Gallop to Southport, Surf's Up! Tomato, good for middle-age problems! lol. Skunkville?